+++ src/heap/spaces.h
@@ -2006,7 +2006,7 @@ class V8_EXPORT_PRIVATE FreeListMany : public FreeList {
   //   say join ", ", @cat;
   //   say "\n", scalar @cat'
   static const int kNumberOfCategories = 46;
-  static constexpr size_t categories_min[kNumberOfCategories] = {
+  static constexpr unsigned int categories_min[kNumberOfCategories] = {
       24,    32,    40,    48,    56,    64,   72,   80,   88,   96,
       104,   112,   120,   128,   136,   144,  152,  160,  168,  176,
       184,   192,   200,   208,   216,   224,  232,  240,  248,  256,
